Update Blue_PDauto.java
Color Sensor works but some measurements need to be updated
VuMark works distance traveled and degree turned needs to be changed
Added Gyro Code but might not work. Needs to be tested
Added CarTurnDegreeDirection(degree, direction) method so that you can input Degree and Direction and robot turns. Needs to be tested.
Added CarSecPow(second, power) method so that you can give second and power and it moved forward or backward for given seconds.